디렉토리 명 변경/ 로그인 response 프로필 추가
