+++ presentation/src/main/java/way/presentation/member/validates/DeleteMemberValidator.java
@@ -1,22 +1,22 @@
-package way.presentation.Member.validates;
+package way.presentation.member.validates;
 
 import org.springframework.stereotype.Component;
 import way.application.utils.exception.BadRequestException;
 import way.application.utils.exception.ErrorResult;
-import way.presentation.Member.vo.req.MemberRequestVo;
+
+import static way.presentation.member.vo.req.MemberRequestVo.*;
 
 @Component
-public class ModifyUserNameValidator {
+public class DeleteMemberValidator {
 
-    public void validate(MemberRequestVo.ModifyUserNameRequest request) {
+    public void validate(DeleteMemberRequest request) {
 
-        validateUserName(request.userName());
         validateMemberSeq(request.memberSeq());
-
+        validateLoginType(request.loginType());
     }
 
-    private void validateUserName(String userName) {
-        if (userName == null || userName.isEmpty()) {
+    private void validateLoginType(String loginType) {
+        if (loginType == null || loginType.isEmpty()) {
             throw new BadRequestException(ErrorResult.DTO_BAD_REQUEST_EXCEPTION);
         }
     }
