+++ presentation/src/main/java/way/presentation/member/validates/LoginValidator.java
@@ -1,26 +1,20 @@
-package way.presentation.Member.validates;
+package way.presentation.member.validates;
 
 import org.springframework.stereotype.Component;
 import way.application.utils.exception.BadRequestException;
 import way.application.utils.exception.ErrorResult;
-import way.presentation.Member.vo.req.MemberRequestVo;
+import way.presentation.member.vo.req.MemberRequestVo;
 
 @Component
-public class SaveSnsMemberValidator {
+public class LoginValidator {
 
-    public void validate(MemberRequestVo.SaveSnsMemberRequest request) {
+    public void validate(MemberRequestVo.LoginRequest request) {
 
-        validateUserName(request.userName());
         validateEmail(request.email());
         validatePassword(request.password());
-        validateLoginType(request.loginType());
+        validateFcmToken(request.fcmToken());
+        validatePassword(request.loginType());
 
-    }
-
-    private void validateUserName(String userName) {
-        if (userName == null || userName.isEmpty()) {
-            throw new BadRequestException(ErrorResult.DTO_BAD_REQUEST_EXCEPTION);
-        }
     }
 
     private void validateEmail(String email) {
@@ -31,6 +25,12 @@ public class SaveSnsMemberValidator {
 
     private void validatePassword(String password) {
         if (password == null || password.isEmpty()) {
+            throw new BadRequestException(ErrorResult.DTO_BAD_REQUEST_EXCEPTION);
+        }
+    }
+
+    private void validateFcmToken(String fcmToken) {
+        if (fcmToken == null || fcmToken.isEmpty()) {
             throw new BadRequestException(ErrorResult.DTO_BAD_REQUEST_EXCEPTION);
         }
     }
