+++ presentation/src/main/java/way/presentation/member/validates/ModifyUserNameValidator.java
@@ -1,23 +1,22 @@
-package way.presentation.Member.validates;
+package way.presentation.member.validates;
 
 import org.springframework.stereotype.Component;
 import way.application.utils.exception.BadRequestException;
 import way.application.utils.exception.ErrorResult;
-import way.presentation.Member.vo.req.MemberRequestVo;
-
-import static way.presentation.Member.vo.req.MemberRequestVo.*;
+import way.presentation.member.vo.req.MemberRequestVo;
 
 @Component
-public class DeleteMemberValidator {
+public class ModifyUserNameValidator {
 
-    public void validate(DeleteMemberRequest request) {
+    public void validate(MemberRequestVo.ModifyUserNameRequest request) {
 
+        validateUserName(request.userName());
         validateMemberSeq(request.memberSeq());
-        validateLoginType(request.loginType());
+
     }
 
-    private void validateLoginType(String loginType) {
-        if (loginType == null || loginType.isEmpty()) {
+    private void validateUserName(String userName) {
+        if (userName == null || userName.isEmpty()) {
             throw new BadRequestException(ErrorResult.DTO_BAD_REQUEST_EXCEPTION);
         }
     }
