+++ presentation/src/main/java/way/presentation/member/validates/SendEmailRequestValidator.java
@@ -1,11 +1,10 @@
-package way.presentation.Member.validates;
+package way.presentation.member.validates;
 
 import org.springframework.stereotype.Component;
 import way.application.utils.exception.BadRequestException;
 import way.application.utils.exception.ErrorResult;
-import way.presentation.Member.vo.req.MemberRequestVo;
 
-import static way.presentation.Member.vo.req.MemberRequestVo.*;
+import static way.presentation.member.vo.req.MemberRequestVo.*;
 
 @Component
 public class SendEmailRequestValidator {
