+++ presentation/src/main/java/way/presentation/member/vo/res/MemberResponseVo.java
@@ -1,4 +1,4 @@
-package way.presentation.Member.vo.res;
+package way.presentation.member.vo.res;
 
 import java.util.List;
